Refactor PowerPlantData to utils script
